+++ corehq/ex-submodules/casexml/apps/case/sharedmodels.py
@@ -1,4 +1,6 @@
 import mimetypes
+
+from corehq.form_processor.abstract_models import CaseAttachmentMixin
 from dimagi.ext.couchdbkit import StringProperty, IntegerProperty, DictProperty
 
 from dimagi.utils.mixins import UnicodeMixIn
@@ -48,7 +50,7 @@ class CommCareCaseIndex(LooselyEqualDocumentSchema, UnicodeMixIn):
         return str(self)
 
 
-class CommCareCaseAttachment(LooselyEqualDocumentSchema, UnicodeMixIn):
+class CommCareCaseAttachment(LooselyEqualDocumentSchema, CaseAttachmentMixin, UnicodeMixIn):
     identifier = StringProperty()
     attachment_src = StringProperty()
     attachment_from = StringProperty()
@@ -58,26 +60,6 @@ class CommCareCaseAttachment(LooselyEqualDocumentSchema, UnicodeMixIn):
 
     attachment_size = IntegerProperty()  # file size
     attachment_properties = DictProperty()  # width, height, other relevant metadata
-
-    @property
-    def is_image(self):
-        if self.server_mime is None:
-            return None
-        return True if self.server_mime.startswith('image/') else False
-
-    @property
-    def is_present(self):
-        """
-        Helper method to see if this is a delete vs. update
-        """
-        if self.identifier and (self.attachment_src == self.attachment_from is None):
-            return False
-        else:
-            return True
-
-    @property
-    def attachment_key(self):
-        return self.identifier
 
     @classmethod
     def from_case_index_update(cls, attachment):
